Deparser: Rework comment handling for PG 18 multi-statement strings
Previously a subsequent statement in a multi-statement string would start
at the end of the prior statement, including any comments between. In
Postgres 18 this has changed, and such comments (and whitespace) are
excluded from the statements, the next statement starts at the first
non-whitespace (and non-comment) character.

Because of this we change the comments logic in the deparser: Previously
comments that had a location outside of the statement would be ignored
by deparseRawStmtOpts, which processes a single RawStmt. Instead, ensure
that all comments passed by the caller are emitted. This is a backwards
incompatible change for any direct callers of the deparser code.

For callers that call libpg_query instead of the deparser directly, e.g.
by using pg_query_deparse_protobuf_opts, nothing changes, as comments
are now split up before passing them down to the deparser code.