Rework PL/pgSQL parsing to use proper type definitions
This is a substantial change to the hack-ish approach we've taken in
prior versions, instead of hacking our way to get the PL/pgSQL parser
to accept our constructs, imitate actual CREATE FUNCTION logic and
PL/pgSQL compilation for a function.

As part of a new release this now requires calling the
scripts/extract_pg_types.rb function to refresh built-in Postgres types.

In passing this improves the mock handling in the extract_source script.